refactor: AV1 sample entry生成処理を共通化
