+++ src/lib.rs
@@ -55,6 +55,7 @@ pub mod subcommand_tune;
 pub mod subcommand_vmaf;
 pub mod types;
 pub mod video;
+pub mod video_av1;
 pub mod video_h264;
 pub mod video_h265;
 pub mod writer_mp4;
